Se reubican funciones a fonts.h y fonts.c
