+++ src/apps/alpha/alpha.c
@@ -4,31 +4,6 @@
 Button sw2,sw3;
 
 uint8_t counter=0;
-uint8_t word[2][8]={0};
-
-int get_char(char c){
-    if(c>='A' && c<='Z'){
-        return c-'A';
-    }
-    else if(c>='a' && c<='z'){
-        return c-'A'-6;
-    }
-    else if(c>='0' && c<='9'){
-        return c + 4;
-    }
-    else{
-        switch(c){
-            case ' ': return 62;
-            case '!': return 63;
-            case '?': return 64;
-            case '/': return 65;
-            case '.': return 66;
-            case ':': return 67;
-            case ';': return 68;
-            default: return -1; //caracter no soportado
-        }
-    }
-}
 
 void load_char_to_screen(char* c){
     Matrix_t *m = GetMatrix();
@@ -43,31 +18,7 @@ void load_char_to_screen(char* c){
     }
 }
 
-void scroll_text(char* text){
-    int location=0;
-    Matrix_t *m = GetMatrix();
-    size_t len=strlen(text);
-    for(int i=0; i<len; i++){
-        location=get_char(text[i]);
-        if(location!=-1){
-            for(int j=0; j<8; j++){
-                word[1][j] = letters[location][j];
-            }
-            for(int j=0; j<8; j++){
-                load_output(m,word[0]);
-                for(int k=0; k<8; k++){
-                    word[0][k] = (word[0][k] << 1);
-                    if((1<<7) & word[1][k]){
-                        word[0][k]++;
-                    }
-                    word[1][k] = (word[1][k] << 1);
-                }
-                shift_matrix(m,0);
-                vTaskDelay(pdMS_TO_TICKS(100));
-            }
-        }
-    }
-}
+
 
 void ButtonHandler(const Button *btn, ButtonEvent_t event)
 {
