+++ src/fonts.h
@@ -6,5 +6,8 @@
 
 extern uint8_t digits[10][8];
 extern uint8_t letters[69][8];
-         
+
+int get_char(char c);
+void scroll_text(char* text);
+
 #endif
